requirements.txt
torch>=1.9.0
transformers>=4.11.0
datasets>=1.11.0
pytorch-lightning>=2.4.0
wandb>=0.12.0
hydra-core>=1.1.0
omegaconf>=2.1.1
hydra-ray-launcher>=1.2.1